moved feeds and entries their own class files
